Fix infinite requests to Thoas on broken network
